net/netutil: unify two oneConnListeners into a new package
I was about to add a third copy, so unify them now instead.
